+++ changelog.md
@@ -10,6 +10,7 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
 
 ### Added
 - NPM `package-lock.json` file.
+- Support for Monster RSS feed and Jobs2Careers.
 
 ### Changed
 - Switched Jobs Hub advert out for new email to upgrade link.